+++ CHANGELOG.md
@@ -9,3 +9,10 @@
 ## [1.0.3] - 2019-12-22
 ### Added
 - "shorten" method.
+
+## [1.0.4] - 2019-12-23
+### Added
+- "roundN" method.
+
+### Fixed
+- An issue with the shorten method.